Move duplicated query type extraction to query utils
