+++ packages/schema/parse/src/extract/query-types.ts
@@ -1,16 +1,14 @@
 import {
   TypeInfo,
-  PropertyDefinition,
-  QueryDefinition,
-  MethodDefinition,
   createQueryDefinition,
   createMethodDefinition,
-  createPropertyDefinition,
-  createScalarDefinition,
-  createArrayDefinition,
-  createObjectDefinition,
 } from "../typeInfo";
-import { getObjectDefinition } from "./utils";
+import {
+  extractInputValueDefinition,
+  extractListType,
+  extractNamedType,
+  State,
+} from "./query-utils";
 
 import {
   DocumentNode,
@@ -22,14 +20,6 @@ import {
   InputValueDefinitionNode,
   visit,
 } from "graphql";
-
-interface State {
-  currentQuery?: QueryDefinition;
-  currentMethod?: MethodDefinition;
-  currentArgument?: PropertyDefinition;
-  currentReturn?: PropertyDefinition;
-  nonNullType?: boolean;
-}
 
 const visitorEnter = (typeInfo: TypeInfo, state: State) => ({
   ObjectTypeDefinition: (node: ObjectTypeDefinitionNode) => {
@@ -60,113 +50,16 @@ const visitorEnter = (typeInfo: TypeInfo, state: State) => ({
     state.currentMethod = method;
   },
   InputValueDefinition: (node: InputValueDefinitionNode) => {
-    const method = state.currentMethod;
-
-    if (!method) {
-      return;
-    }
-
-    const argument = createPropertyDefinition(node.name.value);
-    method.arguments.push(argument);
-    state.currentArgument = argument;
+    extractInputValueDefinition(node, state);
   },
   NonNullType: (_node: NonNullTypeNode) => {
     state.nonNullType = true;
   },
   NamedType: (node: NamedTypeNode) => {
-    const argument = state.currentArgument;
-    const method = state.currentMethod;
-    const modifier = state.nonNullType ? "" : "?";
-
-    if (method && argument) {
-      // Argument value
-      const objectDefinition = getObjectDefinition(
-        node.name.value,
-        typeInfo.userTypes.concat(typeInfo.importedObjectTypes)
-      );
-      if (objectDefinition) {
-        const type = modifier + node.name.value;
-        argument.object = createObjectDefinition(
-          argument.name,
-          type,
-          state.nonNullType,
-          objectDefinition.properties
-        );
-        argument.type = type;
-      } else {
-        argument.scalar = createScalarDefinition(
-          argument.name,
-          modifier + node.name.value,
-          state.nonNullType
-        );
-      }
-
-      state.nonNullType = false;
-    } else if (method) {
-      // Return value
-      if (!method.return) {
-        method.return = createPropertyDefinition(method.name);
-        state.currentReturn = method.return;
-      } else if (!state.currentReturn) {
-        state.currentReturn = method.return;
-      }
-
-      const objectDefinition = getObjectDefinition(
-        node.name.value,
-        typeInfo.userTypes.concat(typeInfo.importedObjectTypes)
-      );
-      if (objectDefinition) {
-        const type = modifier + node.name.value;
-        state.currentReturn.object = createObjectDefinition(
-          method.name,
-          type,
-          state.nonNullType,
-          objectDefinition.properties
-        );
-        state.currentReturn.required = state.nonNullType
-          ? state.nonNullType
-          : null;
-        state.currentReturn.type = type;
-      } else {
-        state.currentReturn.scalar = createScalarDefinition(
-          method.name,
-          modifier + node.name.value,
-          state.nonNullType
-        );
-      }
-      state.nonNullType = false;
-    }
+    extractNamedType(node, state, typeInfo);
   },
   ListType: (_node: ListTypeNode) => {
-    const argument = state.currentArgument;
-    const method = state.currentMethod;
-
-    if (method && argument) {
-      // Argument value
-      argument.array = createArrayDefinition(
-        argument.name,
-        "TBD",
-        state.nonNullType
-      );
-      state.currentArgument = argument.array;
-      state.nonNullType = false;
-    } else if (method) {
-      // Return value
-      if (!method.return) {
-        method.return = createPropertyDefinition(method.name);
-        state.currentReturn = method.return;
-      } else if (!state.currentReturn) {
-        state.currentReturn = method.return;
-      }
-
-      state.currentReturn.array = createArrayDefinition(
-        method.name,
-        "TBD",
-        state.nonNullType
-      );
-      state.currentReturn = state.currentReturn.array;
-      state.nonNullType = false;
-    }
+    extractListType(state);
   },
 });
