--- packages/schema/parse/src/extract/utils.ts
@@ -1,12 +1,0 @@
-import { ObjectDefinition } from "../typeInfo";
-
-export function getObjectDefinition(
-  value: string,
-  objects: ObjectDefinition[]
-): ObjectDefinition | void {
-  for (const object of objects) {
-    if (object.name === value) {
-      return object;
-    }
-  }
-}
